Add graph and tool def models and mapping model
The added models work for generating an editor that can do everything
except create problems. This requires a change to the metamodel (which
also makes sense otherwise): problems will be contained in the Contest
and only referenced by LanguageToProblemMap instances, this way a
problem description can be reused for several languages (for example
in case not all problem descriptions have been translated).